+++ NEWS.md
@@ -6,7 +6,8 @@ v0.2.2 (201?-??-??)
 
 ### New Features ###
 
-* New method for `measurement` to parse a string into a `Measurement` object.
+* New method for `measurement`: you can parse a string in order to create a
+  `Measurement` object.
 * It is now possible to perform calculations involving numbers that have `NaN`
   as nominal value and/or uncertainty.  In previous versions it was possible
   only to define such numbers, but not to perform calculations with them.